Implement more instructions, add stubs for a bunch of subsystems, start timer impl

+++ src/gd/twohundred/jvb/Screen.java
@@ -3,6 +3,6 @@ package gd.twohundred.jvb;
 public interface Screen {
     int WIDTH = 384;
     int HEIGHT = 224;
-    double DISPLAY_REFRESH_RATE_HZ = 50.0;
+    long DISPLAY_REFRESH_RATE_HZ = 50;
     void update(RenderedFrame left, RenderedFrame right);
 }
